Add files via upload
In this project, I explore if male and female have a similar likely hood of a heart attack and if there is a difference between the mean age of male and female when they are at a high risk of heart disease. This report was for a Applied Analytics module and I have used visualization and statistics to analyze and showcase what I have learnt.